refactor: Refatoração no Encriptador de Senhas para Aprimoramento. Moviddo para infraestructura por melhor segurança.
