+++ src/Backend/MyRecipeBook.Application/DependencyInjectionExtension.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using MyRecipeBook.Application.Services.AutoMapper;
-using MyRecipeBook.Application.Services.Cryptografhy;
 using MyRecipeBook.Application.UseCases.Login.DoLogin;
 using MyRecipeBook.Application.UseCases.User.Profile;
 using MyRecipeBook.Application.UseCases.User.Register;
@@ -12,7 +11,6 @@ public static class DependencyInjectionExtension
 {
     public static void AddAplication(this IServiceCollection services, IConfiguration configuration)
     {
-        AddPasswordsEncrpter(services, configuration);
         AddAutoMapper(services);
         AddUseCases(services);
     }
@@ -31,12 +29,5 @@ public static class DependencyInjectionExtension
         services.AddScoped<IDoLoginUseCase, DoLoginUseCase>();
         services.AddScoped<IGetUserProfileUseCase, GetUserProfileUseCase>();
         services.AddScoped<IUpdateUserUseCase, UpdateUserUseCase>();
-    }
-
-    private static void AddPasswordsEncrpter(IServiceCollection services, IConfiguration configuration)
-    {
-        var additionalKey = configuration.GetValue<string>("Settings:Password:AdditionalKey");
-
-        services.AddScoped(option => new PasswordEncripter(additionalKey!));
     }
 }
