+++ src/Backend/MyRecipeBook.Application/UseCases/Login/DoLogin/DoLoginUseCase.cs
@@ -1,7 +1,7 @@
-﻿using MyRecipeBook.Application.Services.Cryptografhy;
-using MyRecipeBook.Communication.Request;
+﻿using MyRecipeBook.Communication.Request;
 using MyRecipeBook.Communication.Responses;
 using MyRecipeBook.Domain.Repositories.User;
+using MyRecipeBook.Domain.Security.Cryptography;
 using MyRecipeBook.Domain.Security.Tokens;
 using MyRecipeBook.Exceptions.ExceptionsBase;
 
@@ -9,10 +9,10 @@ namespace MyRecipeBook.Application.UseCases.Login.DoLogin;
 public class DoLoginUseCase : IDoLoginUseCase
 {
     private readonly IUserReadOnlyRepository _repository;
-    private readonly PasswordEncripter _passwordEncripter;
+    private readonly IPasswordEncripter _passwordEncripter;
     private readonly IAccessTokenGenerator _accessTokenGenerator;
 
-    public DoLoginUseCase(IUserReadOnlyRepository repository, PasswordEncripter passwordEncripter, IAccessTokenGenerator accessTokenGenerator)
+    public DoLoginUseCase(IUserReadOnlyRepository repository, IPasswordEncripter passwordEncripter, IAccessTokenGenerator accessTokenGenerator)
     {
         _repository = repository;
         _passwordEncripter = passwordEncripter;
