+++ src/Backend/MyRecipeBook.Domain/Security/Cryptography/IPasswordEncripter.cs
@@ -1,0 +1,5 @@
+﻿namespace MyRecipeBook.Domain.Security.Cryptography;
+public interface IPasswordEncripter
+{
+    public string Encrypt(string password);
+}
