+++ src/Backend/MyRecipeBook.Infrastucture/DependencyInjectionExtension.cs
@@ -4,11 +4,13 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using MyRecipeBook.Domain.Repositories;
 using MyRecipeBook.Domain.Repositories.User;
+using MyRecipeBook.Domain.Security.Cryptography;
 using MyRecipeBook.Domain.Security.Tokens;
 using MyRecipeBook.Domain.Services.LoggedUser;
 using MyRecipeBook.Infrastucture.DataAccess;
 using MyRecipeBook.Infrastucture.DataAccess.Repositories;
 using MyRecipeBook.Infrastucture.Extensions;
+using MyRecipeBook.Infrastucture.Security.Cryptography;
 using MyRecipeBook.Infrastucture.Security.Tokens.Access.Generator;
 using MyRecipeBook.Infrastucture.Security.Tokens.Access.Validator;
 using MyRecipeBook.Infrastucture.Services.LoggedUser;
@@ -19,6 +21,7 @@ public static class DependencyInjectionExtension
 {
     public static void AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
     {
+        AddPasswordEncrpter(services, configuration);
         AddRepositories(services);
         AddLoggedUser(services);
         AddTokens(services, configuration);
@@ -71,4 +74,11 @@ public static class DependencyInjectionExtension
     }
 
     private static void AddLoggedUser(IServiceCollection services) => services.AddScoped<ILoggedUser, LoggedUser>();
+
+    private static void AddPasswordEncrpter(IServiceCollection services, IConfiguration configuration)
+    {
+        var additionalKey = configuration.GetValue<string>("Settings:Password:AdditionalKey");
+
+        services.AddScoped<IPasswordEncripter>(option => new Sha512Encripter(additionalKey!));
+    }
 }
