+++ src/Backend/MyRecipeBook.Infrastucture/Security/Cryptography/Sha512Encripter.cs
@@ -1,13 +1,13 @@
-﻿using System.Security.Cryptography;
+﻿using MyRecipeBook.Domain.Security.Cryptography;
+using System.Security.Cryptography;
 using System.Text;
 
-namespace MyRecipeBook.Application.Services.Cryptografhy;
-
-public class PasswordEncripter
+namespace MyRecipeBook.Infrastucture.Security.Cryptography;
+public class Sha512Encripter : IPasswordEncripter
 {
     private readonly string _additionalKey;
 
-    public PasswordEncripter(string additionalKey) => _additionalKey = additionalKey;
+    public Sha512Encripter(string additionalKey) => _additionalKey = additionalKey;
 
     public string Encrypt(string password)
     {
