+++ tests/UseCases.Test/User/Register/RegisterUserUseCaseTest.cs
@@ -1,4 +1,4 @@
-﻿using CommonTestUtilities.Cryptografhy;
+﻿using CommonTestUtilities.Cryptography;
 using CommonTestUtilities.Mapper;
 using CommonTestUtilities.Repositories;
 using CommonTestUtilities.Requests;
